+++ src/realm/object-store/c_api/list.cpp
@@ -73,6 +73,29 @@ RLM_API bool realm_list_set(realm_list_t* list, size_t index, realm_value_t valu
     });
 }
 
+RLM_API realm_object_t* realm_list_insert_embedded(realm_list_t* list, size_t index)
+{
+    return wrap_err([&]() {
+        return new realm_object_t({list->get_realm(), list->insert_embedded(index)});
+    });
+}
+
+RLM_API realm_object_t* realm_list_set_embedded(realm_list_t* list, size_t index)
+{
+    return wrap_err([&]() {
+        list->verify_attached();
+        return new realm_object_t({list->get_realm(), list->set_embedded(index)});
+    });
+}
+
+RLM_API realm_object_t* realm_list_get_linked_object(realm_list_t* list, size_t index)
+{
+    return wrap_err([&]() {
+        list->verify_attached();
+        return new realm_object_t({list->get_realm(), list->get_object(index)});
+    });
+}
+
 RLM_API bool realm_list_erase(realm_list_t* list, size_t index)
 {
     return wrap_err([&]() {